Added logging and unit tests for ProductService class

+++ src/main/java/com/example/store_management_tool/services/ProductService.java
@@ -3,6 +3,8 @@ package com.example.store_management_tool.services;
 import com.example.store_management_tool.data.entities.Product;
 import com.example.store_management_tool.data.ProductRequestDto;
 import com.example.store_management_tool.data.repositories.ProductRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,7 +17,9 @@ import java.util.List;
 @Service
 public class ProductService {
     @Autowired
-    public ProductRepository repository;
+    private ProductRepository repository;
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ProductService.class);
 
     public ResponseEntity<List<Product>> getAllProducts() {
         try
@@ -24,7 +28,7 @@ public class ProductService {
         }
         catch (Exception e)
         {
-            e.printStackTrace();
+            LOGGER.error(e.getMessage());
         }
 
         return new ResponseEntity<>(new ArrayList<>(), HttpStatus.BAD_REQUEST);
@@ -33,34 +37,71 @@ public class ProductService {
     public ResponseEntity<Product> getProductById(int id) {
         try
         {
+            if (id <= 0) {
+                LOGGER.error("Invalid ID provided for retrieving product operation");
+                return new ResponseEntity<>(new Product(), HttpStatus.BAD_REQUEST);
+            }
+
             if (repository.findById(id).isPresent())
                 return new ResponseEntity<>(repository.findById(id).get(), HttpStatus.OK);
         }
         catch (Exception e)
         {
-            e.printStackTrace();
+            LOGGER.error(e.getMessage());
         }
 
         return new ResponseEntity<>(new Product(), HttpStatus.NOT_FOUND);
     }
 
     public ResponseEntity<String> addProduct(ProductRequestDto productDto) {
-        Product productToBeAdded = new Product();
+        try
+        {
+            if (productDto.getName().matches(".*\\d.*"))
+            {
+                LOGGER.error("Invalid product name provided");
+                return new ResponseEntity<>("Product name should not contain any digit", HttpStatus.BAD_REQUEST);
+            }
 
-        if (repository.findByName(productDto.getName()).isPresent()) {
-            productToBeAdded = repository.findByName(productDto.getName()).get();
-            productToBeAdded.setQuantity(productToBeAdded.getQuantity() + productDto.getQuantity());
+            if (productDto.getQuantity() <= 0){
+                LOGGER.error("Invalid quantity provided");
+                return new ResponseEntity<>("Product quantity must be greater than 0", HttpStatus.BAD_REQUEST);
+            }
+
+            Product productToBeAdded = new Product();
+
+            if (repository.findByName(productDto.getName()).isPresent()) {
+                LOGGER.error("Product already present, updating the quantity");
+                productToBeAdded = repository.findByName(productDto.getName()).get();
+                productToBeAdded.setQuantity(productToBeAdded.getQuantity() + productDto.getQuantity());
+            } else {
+                BeanUtils.copyProperties(productDto, productToBeAdded);
+            }
+
+            repository.save(productToBeAdded);
         }
-        else {
-            BeanUtils.copyProperties(productDto, productToBeAdded);
+        catch (Exception e)
+        {
+            LOGGER.error(e.getMessage());
         }
 
-        repository.save(productToBeAdded);
         return new ResponseEntity<>("Product added successfully", HttpStatus.CREATED);
     }
 
-    public ResponseEntity<String> deleteProduct(int id){
-        repository.deleteById(id);
+    public ResponseEntity<String> deleteProduct(int id) {
+        try
+        {
+            if (id <= 0) {
+                LOGGER.error("Invalid ID provided for deleting product operation");
+                return new ResponseEntity<>("Invalid ID", HttpStatus.BAD_REQUEST);
+            }
+
+            repository.deleteById(id);
+        }
+        catch (Exception e)
+        {
+            LOGGER.error(e.getMessage());
+        }
+
         return new ResponseEntity<>("Product deleted successfully", HttpStatus.OK);
     }
 }
